BackEnd/GenerateImage/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
diffusers==0.31.0
torch==2.5.0+cu124  
requests==2.32.3
pyjwt==2.10.1